+++ 12_Kultin/Class1/Project1/source.cpp
@@ -1,0 +1,11 @@
+#include "Rac.h";
+
+using namespace std;
+
+void main(){
+	Rac t1, t2;
+	t1.get(2, 13);
+	t2.get(1, 2);
+	Rac().algAct(t1, t2, "/");
+
+};
