+++ 12_Kultin/arr2/13/soucre.c
