+++ 12_Kultin/arr2/19/source.c
