+++ 12_Kultin/testArrs/1arr2/AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.10.35122.118
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA", "AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA.vcxproj", "{4E62D8D4-9DB4-4D21-AD0E-A3BB5862A66E}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{4E62D8D4-9DB4-4D21-AD0E-A3BB5862A66E}.Debug|x64.ActiveCfg = Debug|x64
+		{4E62D8D4-9DB4-4D21-AD0E-A3BB5862A66E}.Debug|x64.Build.0 = Debug|x64
+		{4E62D8D4-9DB4-4D21-AD0E-A3BB5862A66E}.Debug|x86.ActiveCfg = Debug|Win32
+		{4E62D8D4-9DB4-4D21-AD0E-A3BB5862A66E}.Debug|x86.Build.0 = Debug|Win32
+		{4E62D8D4-9DB4-4D21-AD0E-A3BB5862A66E}.Release|x64.ActiveCfg = Release|x64
+		{4E62D8D4-9DB4-4D21-AD0E-A3BB5862A66E}.Release|x64.Build.0 = Release|x64
+		{4E62D8D4-9DB4-4D21-AD0E-A3BB5862A66E}.Release|x86.ActiveCfg = Release|Win32
+		{4E62D8D4-9DB4-4D21-AD0E-A3BB5862A66E}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {9AEFBF22-4AAE-4F6D-8FF9-BEE428B26A03}
+	EndGlobalSection
+EndGlobal
