+++ 12_Kultin/testArrs/1arr2/source.c
