+++ 12_Kultin/testArrs/2arr1/source.c
