--- 12_Культин/Class1/Project1/Rac.h
@@ -1,12 +1,0 @@
-#pragma once
-
-class Rac {
-    private:
-        int chis;
-        int znam;
-    public:
-        void plus(Rac s1,Rac s2);
-        // minus(Rac v1, v2);
-        // umn(Rac m1, Rac m2);
-        // del(Rac d1; Rac d2);
-};
