--- 12_Культин/Class1/Project1/source.cpp
@@ -1,9 +1,0 @@
-#include "Rac.h";
-
-using namespace std;
-
-void main(){
-	Rac t1, t2;
-
-
-};
