--- 12_Культин/Teylor/AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA.vcxproj.user
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="Current" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup />
-</Project>
